ai_services/requirements.txt
# Core FastAPI app
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2

# Google Gemini SDK
google-generativeai==0.3.2

# MongoDB
pymongo==4.6.0

# For file/image handling
pillow==10.1.0

# RAG: Vector DB and retrieval
chromadb==0.4.22

# NLP: Embeddings + retrieval
sentence-transformers==2.2.2
transformers==4.31.0
huggingface_hub==0.16.4
torch>=1.12.1

# Machine Learning (optional for custom similarity or clustering)
numpy>=1.21.0
scikit-learn>=1.0.0

# Optional cleanup
optree>=0.13.0  # To silence PyTorch warnings
